+++ src/app/Constants.cs
@@ -7,11 +7,12 @@
     {
         public static readonly string SwaggerVersion = "v " + Version.AssemblyVersion;
         public const string SwaggerName = "helium";
-        public const string SwaggerTitle = "Helium (C#)";
-        public const string SwaggerPath = "/swagger/" + SwaggerName + "/swagger.json";
+        public const string SwaggerTitle = "Helium";
+        public const string SwaggerPath = "/swagger.json";
         public const string XmlCommentsPath = SwaggerName + ".xml";
 
         public const string KeyVaultName = "KeyVaultName";
+        public const string AuthType = "AUTH_TYPE";
 
         public const string CosmosCollection = "CosmosCollection";
         public const string CosmosDatabase = "CosmosDatabase";